// src/app/model/user.ts
export class User {
    DNI: string;
    nombre: string;
    nombreUsuario: string;
    email: string;
    apellido: string;
    fechaNacimiento: string;
    telefono: string;
    pais: string;
    provincia: string;
    ciudad: string;
    codigoPostal: string;
    direccion: string;
    estadoCuenta: string;
    fechaRegistro: string;
    idiomaPreferido: string;
    fotoPerfil: string;

    constructor(){
        this.DNI = '';
        this.nombre = '';
        this.nombreUsuario = '';
        this.email = '';
        this.apellido = '';
        this.fechaNacimiento = '';
        this.telefono = '';
        this.pais = '';
        this.provincia = '';
        this.ciudad = '';
        this.codigoPostal = '';
        this.direccion = '';
        this.estadoCuenta = '';
        this.fechaRegistro = '';
        this.idiomaPreferido = '';
        this.fotoPerfil = '';
    }
}